new inheritance based tool reformating
