+++ sentinel_1/s1_main.py
@@ -6,7 +6,19 @@ from sentinel_1.denoiser import Denoiser
 from sentinel_1.snap_converter import SnapPreprocessor
 from sentinel_1.tool_manager import ToolManager
 
-from sentinel_1.tool_manager_inheritance import ToolManager as TESTMANAGER
+from sentinel_1.tools.align_raster import AlignRaster
+from sentinel_1.tools.change_resolution import ChangeResolution
+from sentinel_1.tools.clip_256 import Clip256
+from sentinel_1.tools.convert_unit import ConvertUnit
+from sentinel_1.tools.copy_dir import CopyDir
+from sentinel_1.tools.land_sea_mask import LandSeaMask
+from sentinel_1.tools.remove_empty import RemoveEmpty
+from sentinel_1.tools.sort_output import SortOutput
+from sentinel_1.tools.split_polarizations import SplitPolarizations
+from sentinel_1.tools.trim_256 import Trim256
+from sentinel_1.tools.warp_crs import WarpCrs
+
+
 
 class S1Preprocessor:
     @property
@@ -46,22 +58,25 @@ class S1Preprocessor:
             self.denoise_modes = [self.denoise_modes]
 
     def s1_workflow(self):
-        geotiff_utils = ToolManager(
-            self.geotiff_dir, "*.tif", threads=self.threads, polarization=self.polarization
-        )
-        # snap_executor = SnapPreprocessor(gpt_path=self.gpt_exe)
-        # denoiser = Denoiser(self.geotiff_dir, self.shape)
-
-        # snap_executor.graph_processing(
-        #     self.safe_dir, self.geotiff_dir, self.pre_process_graph, input_ext=".zip"
+        # geotiff_utils = ToolManager(
+        #     self.geotiff_dir, "*.tif", threads=self.threads, polarization=self.polarization
         # )
+        snap_executor = SnapPreprocessor(gpt_path=self.gpt_exe)
+        denoiser = Denoiser(self.geotiff_dir, self.shape)
 
-        # copy_dir = os.path.join(self.working_dir, "geotiff_copy")
+        snap_executor.graph_processing(
+            self.safe_dir, self.geotiff_dir, self.pre_process_graph, input_ext=".zip"
+        )
+        sys.exit()
+
+        copy_dir = os.path.join(self.working_dir, "geotiff_copy")
+
         # geotiff_utils.util_starter("copy_dir", copy_dir=copy_dir)
         # copy_dir_utils = ToolManager(
         #     copy_dir, "*.tif", threads=1, polarization=self.polarization
         # )
 
+        SplitPolarizations(self.geotiff_dir, self.shape, self.polarization, self.crs).run()
         # geotiff_utils.util_starter(
         #     "split_polarizations",
         #     output_dir=self.geotiff_dir,
@@ -69,46 +84,62 @@ class S1Preprocessor:
         #     crs=self.crs,
         # )
 
-        TEST_utils = TESTMANAGER(
-            self.geotiff_dir, "*.tif", threads=self.threads, polarization=self.polarization
-        )
-        geotiff_utils.util_starter("align_raster")
-        sys.exit()
+        # geotiff_utils.util_starter("align_raster")
+        AlignRaster(input_dir = self.geotiff_dir).run()
 
-        geotiff_utils.util_starter("land_sea_mask", land_sea_mask = self.land_polygon)
-        geotiff_utils.util_starter("remove_empty")
+        # geotiff_utils.util_starter("land_sea_mask", land_sea_mask = self.land_polygon)
+        LandSeaMask(self.geotiff_dir, self.land_polygon).run()
+
+        # geotiff_utils.util_starter("remove_empty")
+        RemoveEmpty(self.geotiff_dir)
 
         for i, denoise_mode in enumerate(self.denoise_modes):
             resolution = 10
-            if not i == 0: copy_dir_utils.util_starter("copy_dir", copy_dir=self.geotiff_dir)
+            # if not i == 0: copy_dir_utils.util_starter("copy_dir", copy_dir=self.geotiff_dir)
 
-            denoiser.select_denoiser(denoise_mode, to_intensity=False)
+            # denoiser.select_denoiser(denoise_mode, to_intensity=False)
 
-            geotiff_utils.util_starter("change_resolution", x_size=resolution)
-            geotiff_utils.util_starter(
-                "convert_unit", source_unit="linear", destination_unit="decibel"
-            )
-            geotiff_utils.util_starter("align_raster")
-            geotiff_utils.util_starter("trimmer_256")
-            geotiff_utils.util_starter(
-                "sort_output",
-                result_dir=self.sentinel_1_output,
-                working_dir=self.working_dir,
-                denoise_mode=denoise_mode,
-                unit="decibel",
-                resolution=resolution,
-            )
+            # geotiff_utils.util_starter("change_resolution", x_size=resolution)
+            ChangeResolution(self.geotiff_dir, resolution).run()
 
-            geotiff_utils.util_starter(
-                "convert_unit", source_unit="decibel", destination_unit="power"
-            )
-            geotiff_utils.util_starter("align_raster")
-            geotiff_utils.util_starter("trimmer_256")
-            geotiff_utils.util_starter(
-                "sort_output",
-                result_dir=self.sentinel_1_output,
-                working_dir=self.working_dir,
-                denoise_mode=denoise_mode,
-                unit="power",
-                resolution=resolution,
-            )
+            # geotiff_utils.util_starter(
+            #     "convert_unit", source_unit="linear", destination_unit="decibel"
+            # )
+            ConvertUnit(self.geotiff_dir, "linear", "decibel").run()
+
+            # geotiff_utils.util_starter("align_raster")
+            AlignRaster(input_dir = self.geotiff_dir).run()
+
+            # geotiff_utils.util_starter("trimmer_256")
+            Trim256(self.geotiff_dir).run()
+
+            # geotiff_utils.util_starter(
+            #     "sort_output",
+            #     result_dir=self.sentinel_1_output,
+            #     working_dir=self.working_dir,
+            #     denoise_mode=denoise_mode,
+            #     unit="decibel",
+            #     resolution=resolution,
+            # )
+            SortOutput(self.geotiff_dir, self.sentinel_1_output, self.working_dir, denoise_mode, "decibel", resolution).run()
+
+            # geotiff_utils.util_starter(
+            #     "convert_unit", source_unit="decibel", destination_unit="power"
+            # )
+            ConvertUnit(self.geotiff_dir, "decibel", "power").run()
+
+            # geotiff_utils.util_starter("align_raster")
+            AlignRaster(input_dir = self.geotiff_dir).run()
+
+            # geotiff_utils.util_starter("trimmer_256")
+            Trim256(self.geotiff_dir).run()
+
+            # geotiff_utils.util_starter(
+            #     "sort_output",
+            #     result_dir=self.sentinel_1_output,
+            #     working_dir=self.working_dir,
+            #     denoise_mode=denoise_mode,
+            #     unit="power",
+            #     resolution=resolution,
+            # )
+            SortOutput(self.geotiff_dir, self.sentinel_1_output, self.working_dir, denoise_mode, "power", resolution).run()
